requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
google-generativeai
python-dotenv
